Make the menu system better
Instead of just loading all the menus all the time, this uses a script
to load menus when needed and free them when not. This should allow for
a much more modular menu system, and it also makes menu design easier,
since menus don't need to implement all of their logic themselves.