Add a connecting screen
Adds a screen which shows when (re)connecting to the board. This is done
with a really jank state machine sort of a thing, and should probably be
reimplemented with a proper finite state machine (would allow for things
like showing whether it is connecting for the first time or
reconnecting), but it works for now so that will be a future project.